smtp-server/requirements.txt
# Core SMTP functionality
aiosmtpd==1.4.4
python-dotenv==1.0.0

# DNS and email validation
dnspython==2.4.2
email-validator==2.1.0
pyspf==2.0.14

# DKIM signing
dkimpy==1.1.5
authres==1.2.0

# Database and queue
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1

# Web interface
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Security
cryptography==41.0.7
bcrypt==4.1.2
pyjwt==2.8.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
asyncpg==0.29.0
click==8.1.7

# Rate limiting
aioredlock==0.7.3

# IP reputation
geoip2==4.8.0
maxminddb==2.5.1